refactor(channels): move turn engine + server loop to core/engine.ts (#162)
Move runTelegramServer + RunTelegramServerInput + processChatMessage +
SttTimeoutError + withTimeout + sleep + abortReasonString out of the
monolith into src/channels/core/engine.ts.

VERBATIM move: the engine body is byte-identical to its former form in
channels/telegram.ts (confirmed by diff). The ONLY edits were import-path
updates for the new depth, including the single inline type import
`import("../config.ts").TelegramAccount` -> `../../config.ts` — explicitly
permitted import-path retargeting, no logic change.

channels/telegram.ts is now a pure barrel re-exporting the full original
public surface. Full suite: 1184 pass / 0 fail (baseline). Typecheck clean.
